requirements.txt
Django
djangorestframework
django-cors-headers
getenv
psycopg2